add auth page

+++ frontend/src/features/database/DeleteWord/model/async/load-database.worker.ts
@@ -1,11 +1,16 @@
 import { call, put } from 'redux-saga/effects'
 import { api } from '@shared/api'
 import { DeleteWordAction } from '../../../../../entities/database'
+import { authSelectors } from '@entities/auth'
+import { appSelect } from '@shared/libs'
 
 export const deleteWordWorker = function* ({ payload }: { payload: { word: string } }): any {
 	try {
 		const { word } = payload
-		yield call(() => api.database.deleteWord({ word }))
+
+		const userId = yield* appSelect(authSelectors.userId)
+
+		yield call(() => api.database.deleteWord(userId, { word }))
 
 		yield put(DeleteWordAction({ data: word }))
 	} catch (e: any) {
